feat(domain-notes--model): flatten module interface
The modules in the model library are deeply nested. Accessing a symbol
from the "access-control-list" module would go something like this
`Entity.AccessControlList.Model`.

With this change we make it shorter by calling
`AccessControlListEntity.Model`. This makes it easier to work with the
modules.

We also changed a few names to make it easier to read and reason about
the code.