build: add morphic-ts package to handle tagged unions
Typescript has tagged unions that allow us to use the "switch" statement
to operate on each possible type in a discriminated union.

While this feature workes great on simple objects, it does not work on
immutablejs records which is the foundational data type for our data
models. Morphic-ts allows us to work with a form on discriminated union
that is type safe and works perfectly with immutablejs records.