feat(domain-notes--model): make "immutable model" export a "Tag"
The "_tag" property is used to discriminate b/w models, particularly
when doing "discriminated union" operations with the "@morphic-ts/adt"
module.

Since this property will be used all over the codebase, and it may be
the case that the name of the property will change later on, we wrap it
in a constant definition and export that instead of having the string
"_tag" littered all over the code base.